+++ plugins/ImageResizer.Plugins.HybridCache/HybridCacheOptions.cs
@@ -5,26 +5,26 @@ namespace ImageResizer.Plugins.HybridCache
     public class HybridCacheOptions
     {
         /// <summary>
-        ///     Where to store the cached files and the database
+        ///     Where to store the cached files and the database (directory path)
         /// </summary>
-        public string DiskCacheDirectory { get; set; }
+        public string CacheLocation { get; set; }
 
         /// <summary>
         ///     How many RAM bytes to use when writing asynchronously to disk before we switch to writing synchronously.
         ///     Defaults to 100MiB.
         /// </summary>
-        public long QueueSizeLimitInBytes { get; set; } = 100 * 1024 * 1024;
+        public long WriteQueueMemoryMb { get; set; } = 100;
 
         /// <summary>
         ///     Defaults to 1 GiB. Don't set below 9MB or no files will be cached, since 9MB is reserved just for empty directory
         ///     entries.
         /// </summary>
-        public long CacheSizeLimitInBytes { get; set; } = 1 * 1024 * 1024 * 1024;
+        public long CacheSizeMb { get; set; } = 1024;
 
         /// <summary>
         ///     The minimum number of bytes to free when running a cleanup task. Defaults to 1MiB;
         /// </summary>
-        public long MinCleanupBytes { get; set; } = 1 * 1024 * 1024;
+        public long EvictionSweepSizeMb { get; set; } = 1;
 
         /// <summary>
         ///     The minimum age of files to delete. Defaults to 10 seconds.
@@ -40,7 +40,7 @@ namespace ImageResizer.Plugins.HybridCache
 
         public HybridCacheOptions(string cacheDir)
         {
-            DiskCacheDirectory = cacheDir;
+            CacheLocation = cacheDir;
         }
     }
 }